Update patch set 2
Patch Set 2:

Nice, but it's old already. dhcpcd-6.4.0 was released yesterday :)

If dhcpcd is controlling SLAAC it will disable the kernel from doing it. The interaction is fine with one exception introduced in 6.4.0 - the Linux kernel ALWAYS addds it's own idea of the link-local address when the interface comes up, unlike BSD where it checks for a LL address before adding one. My testing shows the kernel will only use the first link-local address added to the interface so it's a race between the interface going up and dhcpcd adding the address. This is for RFC7217 support, which the kernel lacks and does need something in userspace for.

As to merging from multiple interface, dhcpcd has been handling this just fine since version 5 and DHCPv6 is no different in this regard.

For reference, Windows has NO support for RDNSS and DNSSL in RA messages so DHCPv6 is the ONLY option to pass this. As such, the most common IPv6 only network setup is to require DHCPv6.

As I'm the dhcpcd author, I have not reviewed this patch as I have obvious bias.

Patch-set: 2
Reviewer: Roy Marples <1048696@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0