Do not IFF_LOWER_UP define check on android.
The latest uapi kernel headers use an enum for values such as
IFF_LOWER_UP. This code defines IFF_LOWER_UP that causes the inclusion
of if.h later on to mess up the enum definition.

Simply include the if.h file before checking for the define rather
than after.

Change-Id: I9d9ce63a63bb6fc82a2424776f5aa6700cc34239