Update patch set 1
Patch Set 1:

I may be confused, but isn't this the same mechanism used to tell us the default gateway and dns servers etc?  This is only run on a network we're connecting to because the user has selected it.  The routes should be reported into android and only affect the routing table of that network..

Patch-set: 1