Update patch set 1
Patch Set 1:

Is there any chance that this could land on a pre KK tree?

Are you assuming that a user would never select a malicious network?

I'm not a big fan of changing the routing behavior by default, but if there is no chance of introducing a serious security problem like a VPN bypass I guess I'm ok with installing this by default.

Patch-set: 1