Update patch set 4
Patch Set 4:

No. dhcpcd has the the code (cf if-linux-wireless.c) to read the current ssid through standard wireless extension linux API (or BSD.). Though, at the moment, it only uses it to send the information to the hooks.
So this change is fully transparent for Android.
There is a change in dumplease, to be able to specify ssid on dumplease's cmdline.

Patch-set: 4