Update patch set 4
Patch Set 4:

Oh, I see. So you're saying that dhcpcd gets the SSID using Linux wireless extensions when dhcpcd is started.

I'm not a fan of that approach. For example, the SSID can change when dhcpcd is running. This happens most commonly when we roam to a different network (e.g., mywireless-2G to mywireless-5G and send SIGALRM to do a renew on the new network), but it can happen in other corner cases (e.g., if the user selects a wireless network in settings and immediately after that selects another network). In that case the lease will be written to the wrong file.

I think it's better if this is handled in the framework.

Patch-set: 4