Update patch set 4
Patch Set 4: Code-Review-1

Not sure if that's sufficient. For example, what happens if:

1. We connect to SSID "network5G", get an IPv4 address, and write the result to the lease file for network5G.
2. We roam to "network2G", and ask dhcpcd to renew.
3. dhcp tries tries to open the leasefile for network2G, but there is no leasefile for it, or the lease has expired.

How do we renew?

Also, looks like this code will not do anything for SSIDs that contain - or _ (or any non-ASCII characters). We'd need to escape those... but if we do, then the framework might escape them in a different way to dhcpcd, and that might cause confusion.

In general I think we're stretching the existing design too far. The problem here is that we have no way to pass the network name and existing IP address to dhcpcd. But there are other problems. For example, every parameter that needs to be passed back from dhcpcd (e.g., static routes) needs to have code added to the hook scripts, JNI parsing code, and Java code. Also, there is no way to get asynchronous notifications from DHCP - all we have is a synchronous, blocking call that can pass down no information except the interface name. And so on.

For this and other reasons, I don't think it's worth putting so much effort into fixing the problem at this low layer. I think it's pretty clear that for this type of thing we need tighter integration between the framework and DHCP, and I think we will be working on that.

Patch-set: 4