Update patch set 4
Patch Set 4:

Pierre (Vandwalle): I expect that the change from 2.5s to 900ms is because when moving to a new network, the DHCP client no longer sends a DHCPREQUEST for IP address it had on the old network.

I suspect a similar performance improvement would be possible by simply blowing away the lease database when switching networks (but that would also cause the DHCP client ID to change).

Patch-set: 4