Update patch set 4
Patch Set 4:

As I said before, I think the way forward here is not to do this work in dhcpcd and wpa_supplicant, but instead work on getting DHCP more integrated with the framework.

Options I see for this are either a rewrite of the dhcp functionality in Java or Java/JNI/C++, or replacing the current Java/JNI/systemproperties-based interface between the framework and dhcpcd with something else (perhaps a FrameworkListener?).

Patch-set: 4