Update patch set 4
Patch Set 4:

"Pierre (Vandwalle): I expect that the change from 2.5s to 900ms is because when moving to a new network, the DHCP client no longer sends a DHCPREQUEST for IP address it had on the old network."

you mean when WiFi disconnect from current WiFi AP, and then reassociate to an AP with new SSID, which means that dhcpcd sees interface going down, and after some time, up with a new SSID?

The framework already tries to manage if DHCP should renew or try to get a new lease, I think as well that this decision should remain there, with tighter integration as Lorenzo mentioned. Also there is no direct relationship in the 802.11 spec between the 802.11 Service Set and IP address management, so having dhcpcd rely on SSID changes doesn't seem to be the right thing to do.

Patch-set: 4
Reviewer: Pierre Vandwalle <1045688@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0