Fixed valid_length()'s error value for IPV4 and ARRAY option types
When parsing DHCP options of type IPV4 or ARRAY, valid_length() would return a positive integer for bad input (as opposed to -1).
The caller checks for -1 return value (for error) and so valid_length() would never return an error for IPV4 or ARRAY option type.
This could have security implications (options of invalid lengths being parsed)

Change-Id: I8f7330f8c527351dd578ea30d1691ec85b683d32
Signed-off-by: Shachar Menashe, NorthBit Ltd <srmish@north-bit.com>